separated program execution from vs code code
